Add ability to delete/update any pin as an admin && added custom error403 page/small changes on error404page

+++ src/Controller/PinsController.php
@@ -8,6 +8,7 @@ use App\Form\PinType;
 use App\Repository\PinRepository;
 use App\Repository\UserRepository;
 use Doctrine\ORM\EntityManagerInterface;
+use Sensio\Bundle\FrameworkExtraBundle\Configuration\IsGranted;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
@@ -30,6 +31,18 @@ class PinsController extends AbstractController
      */
     public function create(Request $request, EntityManagerInterface $em, UserRepository $userRepo): Response
     {
+        if (!$this->getUser()) {
+            $this->addFlash('error', 'You need to log in first!');
+
+            return $this->redirectToRoute('app_login');
+        }
+
+        if (!$this->getUser()->isVerified()) {
+            $this->addFlash('error', 'You need to verify your email first!');
+
+            return $this->redirectToRoute('app_home');
+        }
+
         $pin = new Pin;
 
         $form = $this->createForm(PinType::class, $pin);
@@ -64,6 +77,24 @@ class PinsController extends AbstractController
      */
     public function edit(Request $request, EntityManagerInterface $em, Pin $pin): Response
     {
+        if (!$this->getUser()) {
+            $this->addFlash('error', 'You need to log in first!');
+
+            return $this->redirectToRoute('app_login');
+        }
+
+        if (!$this->getUser()->isVerified()) {
+            $this->addFlash('error', 'You need to verify your email first!');
+
+            return $this->redirectToRoute('app_home');
+        }
+
+        if ($pin->getUser() != $this->getUser() && !$this->IsGranted('ROLE_ADMIN')) {
+            $this->addFlash('error', 'Access forbidden.');
+
+            return $this->redirectToRoute('app_home');
+        }
+
         $form = $this->createForm(PinType::class, $pin);
 
         $form->handleRequest($request);
@@ -90,6 +121,24 @@ class PinsController extends AbstractController
      */
     public function delete(Request $request, EntityManagerInterface $em, Pin $pin): Response
     {
+        if (!$this->getUser()) {
+            $this->addFlash('error', 'You need to log in first!');
+
+            return $this->redirectToRoute('app_login');
+        }
+
+        if (!$this->getUser()->isVerified()) {
+            $this->addFlash('error', 'You need to verify your email first!');
+
+            return $this->redirectToRoute('app_home');
+        }
+
+        if ($pin->getUser() != $this->getUser() && !$this->IsGranted('ROLE_ADMIN')) {
+            $this->addFlash('error', 'Access forbidden.');
+
+            return $this->redirectToRoute('app_home');
+        }
+
         if ($this->isCsrfTokenValid('pin_deletion_' . $pin->getId(), $request->request->get('_csrf_token'))) {
             $em->remove($pin);
             $em->flush();
